initial refactor of pre-processor
