+++ src/app/routes/cpsAsset/getInitialData/convertToOptimoBlocks/blocks/include/ampMetadataExtractor/index.js
@@ -1,4 +1,8 @@
 import Url from 'url-parse';
+import nodeLogger from '#lib/logger.node';
+import { INCLUDE_IFRAME_REQUEST_RECEIVED } from '#lib/logger.const';
+
+const logger = nodeLogger(__filename);
 
 /**
  * Returns an object containing metadata extracted from an Include url
@@ -6,12 +10,19 @@ import Url from 'url-parse';
  * @param {string} href The href value from an Include Block.
  */
 
-const ampMetadataExtractor = (href, baseUrl) => {
+const ampMetadataExtractor = (href, baseUrl, classification) => {
+  if (classification !== 'vj-supports-amp') return null;
   const { hostname, pathname, protocol, query } = new Url(href, baseUrl, true);
   const queryString = new Url(href).query;
 
+  const src = `${protocol}//${hostname}${pathname}/amp${queryString}`;
+
+  logger.info(INCLUDE_IFRAME_REQUEST_RECEIVED, {
+    url: src,
+  });
+
   return {
-    src: `${protocol}//${hostname}${pathname}/amp${queryString}`,
+    src,
     imageWidth: query['amp-image-width'],
     imageHeight: query['amp-image-height'],
     image: query['amp-image'],
