+++ src/app/routes/cpsAsset/getInitialData/convertToOptimoBlocks/blocks/include/getImageBlock/index.js
@@ -1,4 +1,4 @@
-import htmlUnescape from '../../../../../utils/htmlUnescape';
+import htmlUnescape from '../../../../../../utils/htmlUnescape';
 
 const getImageBlock = (type, blockData, isAmp) => {
   const supportedImageTypes = ['idt2'];
