+++ src/app/routes/cpsAsset/getInitialData/convertToOptimoBlocks/blocks/include/includeClassifier/index.test.js
@@ -1,4 +1,4 @@
-import includeClassifier from './includeClassifier';
+import includeClassifier from '.';
 
 const vjIncludeSupportingAmp =
   '/include/newsspec/21841-green-diet/gahuza/app?responsive=true&newsapps=true&app-image=https://news.files.bbci.co.uk/vj/live/idt-images/image-slider-asdf/app_launcher_ws_640_7ania.png&app-clickable=true&amp-clickable=true&amp-image-height=360&amp-image-width=640&amp-image=https://news.files.bbci.co.uk/vj/live/idt-images/image-slider-asdf/app_launcher_ws_640_7ania.png';
@@ -6,9 +6,6 @@ const vjIncludeNotSupportingAmp =
   '/news/special/2016/newsspec_14813/content/iframe/gahuza/us-gop.inc?responsive=true&app-clickable=true&app-image=http://a.files.bbci.co.uk/worldservice/live/assets/images/2016/11/09/161109092836_us_election_2nddaymaps_winner_ws_62_v3.png';
 const idt1Include = '/indepthtoolkit/quizzes/123-456';
 const idt2Include = '/idt2/111-222-333-444-555';
-
-const canonicalPathname = 'https://www.bbc.com/service/foo';
-const ampPathname = 'https://www.bbc.com/service/foo.amp';
 
 describe('Include Classifier', () => {
   it.each`
@@ -26,7 +23,7 @@ describe('Include Classifier', () => {
 
       const actual = includeClassifier({
         href,
-        pathname: canonicalPathname,
+        isAmp: false,
       });
 
       expect(actual).toEqual(expected);
@@ -41,7 +38,7 @@ describe('Include Classifier', () => {
 
     const actual = includeClassifier({
       href: vjIncludeNotSupportingAmp,
-      pathname: canonicalPathname,
+      isAmp: false,
     });
 
     expect(actual).toEqual(expected);
@@ -63,7 +60,7 @@ describe('Include Classifier', () => {
 
       const actual = includeClassifier({
         href,
-        pathname: ampPathname,
+        isAmp: true,
       });
 
       expect(actual).toEqual(expected);
@@ -78,7 +75,7 @@ describe('Include Classifier', () => {
 
     const actual = includeClassifier({
       href: vjIncludeNotSupportingAmp,
-      pathname: ampPathname,
+      isAmp: true,
     });
 
     expect(actual).toEqual(expected);
@@ -92,7 +89,7 @@ describe('Include Classifier', () => {
 
     const actual = includeClassifier({
       href: 'idt3/blah',
-      pathname: canonicalPathname,
+      isAmp: false,
     });
 
     expect(actual).toEqual(expected);
@@ -106,7 +103,7 @@ describe('Include Classifier', () => {
 
     const actual = includeClassifier({
       href: 'idt3/blah',
-      pathname: ampPathname,
+      isAmp: false,
     });
 
     expect(actual).toEqual(expected);
@@ -127,7 +124,7 @@ describe('Include Classifier', () => {
 
       const actual = includeClassifier({
         href,
-        pathname: canonicalPathname,
+        isAmp: false,
       });
 
       expect(actual).toEqual(expected);
@@ -149,7 +146,7 @@ describe('Include Classifier', () => {
 
       const actual = includeClassifier({
         href,
-        pathname: ampPathname,
+        isAmp: true,
       });
 
       expect(actual).toEqual(expected);
