--- src/app/routes/cpsAsset/getInitialData/convertToOptimoBlocks/blocks/include/isAmpSupported.js
@@ -1,6 +1,0 @@
-export default classification =>
-  !(
-    classification === 'vj-amp-not-supported' ||
-    classification === 'idt1-amp' ||
-    classification === 'idt1-canonical'
-  );
