+++ src/app/routes/cpsAsset/getInitialData/convertToOptimoBlocks/blocks/include/utils/index.test.js
@@ -1,15 +1,15 @@
-import isAmpSupported from './isAmpSupported';
-
-const truthyTestCases = [
-  'idt2-canonical',
-  'idt2-amp',
-  'vj-canonical',
-  'vj-supports-amp',
-];
-
-const falsyTestCases = ['vj-amp-not-supported', 'idt1-amp', 'idt1-canonical'];
+import { isAmpSupported } from '.';
 
 describe('isAmpSupported', () => {
+  const truthyTestCases = [
+    'idt2-canonical',
+    'idt2-amp',
+    'vj-canonical',
+    'vj-supports-amp',
+  ];
+
+  const falsyTestCases = ['vj-amp-not-supported', 'idt1-amp', 'idt1-canonical'];
+
   truthyTestCases.forEach(classification => {
     it(`should return true if classification is ${classification}`, () => {
       expect(isAmpSupported(classification)).toEqual(true);
